+++ internal/sandbox/webgamedevimage/embed.go
@@ -1,0 +1,6 @@
+package webgamedevimage
+
+import _ "embed"
+
+//go:embed Dockerfile
+var dockerfileBytes []byte